Updated Functionality
Added sensor readings for force, color, and distance sensors as well as built in accelerometer data in the X, Y, and Z directions. Currently all the nodes have values from 0 to 1. The distance node output is in units of meters, the force node output is the amount pushed down the button is (0 is not at all, 1 is fully pushed down), the accelerometer node output is divided by 5000 because that's about the range of readings. Gravity is therefore around 0.19-0.20 in the node readings. The color node doesn't output anything currently as Vuforia doesn't currently support nodes having strings. Instead, if the color is not "None", it is logged to the server running in terminal (if you aren't running the desktop app). The node stopRead accepts a switch value, and when the switch is turned on, the sensor readings stop, and when the switch is turned back off, sensor readings proceed.